Converted assets to floating point numbers

+++ Engine/Dude.h
@@ -9,15 +9,15 @@ public:
 	void ClampToScreen();
 	void Draw( Graphics& gfx ) const;
 	void Update( const Keyboard& kbd );
-	int GetX() const;
-	int GetY() const;
-	int GetWidth() const;
-	int GetHeight() const;
+	float GetX() const;
+	float GetY() const;
+	float GetWidth() const;
+	float GetHeight() const;
 	void Reset();
 private:
-	int x = 400;
-	int y = 300;
-	static constexpr int speed = 2;
-	static constexpr int width = 20;		//Dimensions of the sprite
-	static constexpr int height = 20;
+	float x = 400.0f;
+	float y = 300.0f;
+	static constexpr float speed = 2.0f;
+	static constexpr float width = 20.0f;		//Dimensions of the sprite
+	static constexpr float height = 20.0f;
 };
