+++ Engine/Poo.h
@@ -6,17 +6,17 @@
 class Poo
 {
 public:
-	Poo( const int _x,const int _y,const int _vx,const int _vy );
+	Poo( const float _x,const float _y,const float _vx,const float _vy );
 	void Update();
 	void TestCollision( const Dude& dude );
 	void Draw( Graphics& gfx ) const;
 	bool IsEaten() const;
-	void Reset(int _x, int _y, int _vx, int _vy);
+	void Reset( float _x,float _y,float _vx,float _vy );
 private:
-	int x;		//x postion
-	int y;		//y position
-	int vx;		//velocity in the x direction
-	int vy;		// velocity in the y direction
+	float x;		//x postion
+	float y;		//y position
+	float vx;		//velocity in the x direction
+	float vy;		// velocity in the y direction
 	static constexpr int width = 24;		//These are the dimensions of the sprite itself
 	static constexpr int height = 24;
 	bool isEaten = false;
